fixes after mentor review
