+++ src/js/main.js
@@ -1,5 +1,5 @@
-import {boardContainer, renderCards} from './modules/_board';
-import renderFilter from './modules/_filter';
+import {renderCards} from './modules/board';
+import {renderFilter} from './modules/filter';
 
 renderCards();
 renderFilter();
@@ -10,7 +10,7 @@ const renderedFilters = document.querySelectorAll(`.filter__label`);
 // Clean Board And Render New Cards
 Array.from(renderedFilters).forEach((renderedFiltersElement) => {
   renderedFiltersElement.addEventListener(`click`, function () {
-    boardContainer.innerHTML = ``;
+    document.querySelector(`.board__tasks`).innerHTML = ``;
     renderCards();
   });
 });
